Render the renew now button if eligible
